refactor: split chat prompt template modules
