+++ test/Test/Langchain/PromptTemplate/Chat/ChatMessagePromptTemplateSpec.hs
@@ -6,14 +6,15 @@ import Test.Tasty
 import Test.Tasty.HUnit
 
 import Langchain.PromptTemplate (PromptTemplate (..))
-import Langchain.PromptTemplate.Chat
+import Langchain.PromptTemplate.Chat (BaseStringMessagePromptTemplate (..))
+import Langchain.PromptTemplate.Chat.ChatMessagePromptTemplate
 
 tests :: TestTree
 tests =
   testGroup
     "ChatMessagePromptTemplate"
-    [ testCase "from_template_file creates a role chat message prompt template" $ do
-        actual <- chatMessagePromptTemplateFromTemplateFile "test/data/prompt_file.txt" "human"
+    [ testCase "fromTemplateFile creates a role chat message prompt template" $ do
+        actual <- fromTemplateFile "test/data/prompt_file.txt" "human" :: IO ChatMessagePromptTemplate
         let expected =
               ChatMessagePromptTemplate
                 { chatMessagePromptTemplatePrompt = PromptTemplate "Question: {question}\nAnswer:"
