+++ test/Test/Langchain/PromptTemplate/Chat/MessagesPlaceholderSpec.hs
@@ -9,31 +9,29 @@ import Test.Tasty.HUnit
 
 import Langchain.Core.Error (errorMessage)
 import Langchain.Core.Model.Types (Role (..), assistantMessage, systemMessage, userMessage)
-import Langchain.PromptTemplate.Chat
+import Langchain.PromptTemplate.Chat (BaseMessagePromptTemplate (..))
+import Langchain.PromptTemplate.Chat.MessagesPlaceholder
 
 tests :: TestTree
 tests =
   testGroup
     "MessagesPlaceholder"
     [ testCase "required placeholder requires its variable" $ do
-        let result = formatMessagesPlaceholder (messagesPlaceholder "history") Map.empty
+        let result = formatMessages (messagesPlaceholder "history") emptyInputs
         case result of
           Left err ->
             "history" `T.isInfixOf` errorMessage err
               @? "Expected error to mention missing history"
           Right _ -> assertFailure "Expected missing history to fail"
     , testCase "optional placeholder formats to an empty list when omitted" $
-        formatMessagesPlaceholder (optionalMessagesPlaceholder "history") Map.empty
+        formatMessages (optionalMessagesPlaceholder "history") emptyInputs
           @?= Right []
     , testCase "optional placeholder converts mixed message representations" $
-        formatMessagesPlaceholder
+        formatMessages
           (optionalMessagesPlaceholder "history")
-          ( Map.fromList
-              [ ( "history"
-                , [ PlaceholderRoleMessage System "You are an AI assistant."
-                  , PlaceholderHumanText "Hello!"
-                  ]
-                )
+          ( inputs
+              [ PlaceholderRoleMessage System "You are an AI assistant."
+              , PlaceholderHumanText "Hello!"
               ]
           )
           @?= Right
@@ -42,9 +40,9 @@ tests =
             ]
     , testCase "placeholder without a message limit returns the whole history" $
         let history = map (PlaceholderMessage . assistantMessage) ["1", "2", "3"]
-         in formatMessagesPlaceholder
+         in formatMessages
               (messagesPlaceholder "history")
-              (Map.fromList [("history", history)])
+              (inputs history)
               @?= Right (map assistantMessage ["1", "2", "3"])
     , testCase "placeholder with n_messages returns the last messages" $
         let history = map (PlaceholderMessage . assistantMessage) ["1", "2", "3"]
@@ -53,8 +51,14 @@ tests =
                 assertFailure $
                   "Expected valid n_messages, got: " <> T.unpack (errorMessage err)
               Right prompt ->
-                formatMessagesPlaceholder
+                formatMessages
                   prompt
-                  (Map.fromList [("history", history)])
+                  (inputs history)
                   @?= Right [assistantMessage "2", assistantMessage "3"]
     ]
+
+emptyInputs :: Map.Map T.Text [MessagePlaceholderInput]
+emptyInputs = Map.empty
+
+inputs :: [MessagePlaceholderInput] -> Map.Map T.Text [MessagePlaceholderInput]
+inputs history = Map.fromList [("history", history)]
